sweep over 2ball_300 revealed that embo curve does not span far enough, because simulation points reach past 6 bits of complexity, curve will probably be similar to wcs meaning space

+++ src/measure.py
@@ -47,7 +47,7 @@ def main(config: DictConfig):
     # Compute matrix of pairwise gNID between emergent and optimal
     gNIDs = torch.tensor(
         [
-            [gNID(em, opt, g.prior) for opt in optimal_encoders]
+            [gNID(em.float(), opt.float(), g.prior.float()) for opt in optimal_encoders]
             for em in emergent_encoders
         ]
     )
